Painter improving

+++ view/character/Painter.h
@@ -23,7 +23,7 @@ public:
 private:
     double initialH;
     double finalH;
-    double offset;
+    int offset;
     Uint32 getpixel(SDL_Surface *surface, int x, int y);
     void putpixel(SDL_Surface *surface, int x, int y, Uint32 pixel);
